+++ src/Components/Button/Base/ButtonBase.tsx
@@ -1,0 +1,94 @@
+// Libraries
+import React, {Component, MouseEvent, RefObject} from 'react'
+import classnames from 'classnames'
+
+// Styles
+import '../Button.scss'
+
+// Types
+import {
+  ComponentStatus,
+  ComponentColor,
+  ComponentSize,
+  ButtonShape,
+  ButtonType,
+} from '../../../Types'
+
+interface Props {
+  /** Function to be called on button click */
+  onClick?: (e?: MouseEvent<HTMLButtonElement>) => void
+  /** Text to be displayed on hover tooltip */
+  titleText?: string
+  /** Keyboard control tab order  */
+  tabIndex?: number
+  /** Class name for custom styles */
+  className?: string
+  /** Button color */
+  color: ComponentColor
+  /** Button size */
+  size: ComponentSize
+  /** Square or rectangle */
+  shape: ButtonShape
+  /** Button status state default, loading, or disabled */
+  status: ComponentStatus
+  /** Toggles button highlighted active state */
+  active: boolean
+  /** Button type of 'button' or 'submit' */
+  type: ButtonType
+  /** Test ID for Integration Tests */
+  testID: string
+}
+
+export class ButtonBase extends Component<Props> {
+  public static defaultProps = {
+    color: ComponentColor.Default,
+    size: ComponentSize.Small,
+    shape: ButtonShape.Default,
+    status: ComponentStatus.Default,
+    active: false,
+    type: ButtonType.Button,
+    testID: ButtonType.Button,
+  }
+
+  public ref: RefObject<HTMLButtonElement> = React.createRef()
+
+  public render() {
+    const {onClick, titleText, tabIndex, type, testID, children} = this.props
+
+    return (
+      <button
+        className={this.className}
+        disabled={this.disabled}
+        onClick={onClick}
+        title={titleText}
+        tabIndex={!!tabIndex ? tabIndex : 0}
+        type={type}
+        ref={this.ref}
+        data-testid={testID}
+      >
+        {children}
+      </button>
+    )
+  }
+
+  private get disabled(): boolean {
+    const {status} = this.props
+
+    return (
+      status === ComponentStatus.Disabled || status === ComponentStatus.Loading
+    )
+  }
+
+  private get className(): string {
+    const {color, size, shape, status, active, className} = this.props
+
+    return classnames(`button button-${size} button-${color}`, {
+      'button-square': shape === ButtonShape.Square,
+      'button-stretch': shape === ButtonShape.StretchToFit,
+      'button--loading': status === ComponentStatus.Loading,
+      'button--disabled': status === ComponentStatus.Disabled,
+      active,
+      [`${className}`]: className,
+    })
+  }
+}