+++ src/Components/Button/Button.stories.tsx
@@ -8,7 +8,9 @@ import {withKnobs, text, select, boolean} from '@storybook/addon-knobs'
 import {mapEnumKeys} from '../../../.storybook/utils'
 
 // Components
-import {Button} from './Button'
+import {Button} from './Composed/Button'
+import {SquareButton} from './Composed/SquareButton'
+import {ButtonBase} from './Base/ButtonBase'
 
 // Types
 import {
@@ -20,11 +22,15 @@ import {
   ButtonType,
 } from '../../Types'
 
-const buttonStories = storiesOf('Components|Buttons', module)
+const buttonBaseStories = storiesOf('Components|Buttons/Base', module)
   .addDecorator(withKnobs)
   .addDecorator(jsxDecorator)
 
-buttonStories.add('Button Component', () => (
+const buttonComposedStories = storiesOf('Components|Buttons/Composed', module)
+  .addDecorator(withKnobs)
+  .addDecorator(jsxDecorator)
+
+buttonComposedStories.add('Standard Button', () => (
   <Button
     text={text('text', 'Button Text')}
     onClick={() => alert('clicked')}
@@ -42,5 +48,42 @@ buttonStories.add('Button Component', () => (
     }
     active={boolean('active', false)}
     type={ButtonType[select('type', mapEnumKeys(ButtonType), 'Button')]}
+    placeIconAfterText={boolean('placeIconAfterText', false)}
   />
+))
+
+buttonComposedStories.add('Square Button (Icon Only)', () => (
+  <SquareButton
+    onClick={() => alert('clicked')}
+    icon={IconFont[select('icon', mapEnumKeys(IconFont), 'Zap')]}
+    titleText={text('titleText', 'Title Text')}
+    color={
+      ComponentColor[select('color', mapEnumKeys(ComponentColor), 'Default')]
+    }
+    size={ComponentSize[select('size', mapEnumKeys(ComponentSize), 'Small')]}
+    status={
+      ComponentStatus[select('status', mapEnumKeys(ComponentStatus), 'Default')]
+    }
+    active={boolean('active', false)}
+    type={ButtonType[select('type', mapEnumKeys(ButtonType), 'Button')]}
+  />
+))
+
+buttonBaseStories.add('Base Button', () => (
+  <ButtonBase
+    onClick={() => alert('clicked')}
+    titleText={text('titleText', 'Title Text')}
+    color={
+      ComponentColor[select('color', mapEnumKeys(ComponentColor), 'Default')]
+    }
+    size={ComponentSize[select('size', mapEnumKeys(ComponentSize), 'Small')]}
+    shape={ButtonShape[select('shape', mapEnumKeys(ButtonShape), 'Default')]}
+    status={
+      ComponentStatus[select('status', mapEnumKeys(ComponentStatus), 'Default')]
+    }
+    active={boolean('active', false)}
+    type={ButtonType[select('type', mapEnumKeys(ButtonType), 'Button')]}
+  >
+    {text('text', 'Button Text')}
+  </ButtonBase>
 ))